Moves pos to src for further refactoring into main app. Refactors out the logging mechanism
